# Adding ibus telemetry support via serial to arducopter