--- a/Tools/ardupilotwaf/ardupilotwaf.py
+++ b/Tools/ardupilotwaf/ardupilotwaf.py
@@ -64,6 +64,7 @@ COMMON_VEHICLE_DEPENDENT_LIBRARIES = [
     'AP_Button',
     'AP_ICEngine',
     'AP_Frsky_Telem',
+    'AP_IBus_Telem',
     'AP_FlashStorage',
     'AP_Relay',
     'AP_ServoRelayEvents',