--- a/libraries/AP_Vehicle/AP_Vehicle.cpp
+++ b/libraries/AP_Vehicle/AP_Vehicle.cpp
@@ -160,7 +160,13 @@ void AP_Vehicle::setup()
 #if GENERATOR_ENABLED
     generator.init();
 #endif
+
+#if HAL_IBUS_TELEM_ENABLED
+    //will try to init IBus telemetry if configured on a serial port
+    ibus_telem.init();
+#endif
 }
+
 
 void AP_Vehicle::loop()
 {
@@ -211,6 +217,9 @@ const AP_Scheduler::Task AP_Vehicle::scheduler_tasks[] = {
 #endif
 #if OSD_ENABLED
     SCHED_TASK(publish_osd_info, 1, 10),
+#endif
+#if HAL_IBUS_TELEM_ENABLED
+    SCHED_TASK_CLASS(AP_IBus_Telem, &vehicle.ibus_telem,    loop,                     400, 85),
 #endif
 };