This is my first pull request. I tried to follow the guidelines as much as possible. 
Sorry if i made newbie mistakes.