@jverhuls 2021-02-07
Dear, it seems that i'm missing something in order to let the build work correctly. I filled in my new library in the wscript, but it seems it can't resolve it. Any idea what i'm missing?