@IamPete1 2021-02-07
I think you don't need it in the individual vehicle wscript files just adding it in ardupilotwaf.py should be fine (with the correct capitalization ;) ).