@jverhuls 2021-02-13
I updated my code with all changes requested, but I messed up the commits during attempt to cleanup. 
What shall I do start from a fresh branch and do my changes there or is there someone who can help me to cleanup my feature branch?