@jverhuls 2021-02-14
> On Sat, 13 Feb 2021, jverhuls wrote: I updated my code with all changes requested, but I messed up the commits during attempt to cleanup. What shall I do start from a fresh branch and do my changes there or is there someone who can help me to cleanup my feature branch?
> I can fix it up for you if you like.

That's very much appreciated. Is it possible to tell me how you can fix it, so I can learn?