@peterbarker 2021-02-14
@jverhuls   The git history was basically to messy to be worth cleaning up, so I essentially just copied all of your changes into a new branch:

```
pbarker@bluebottle:~/rc/ardupilot(master)$ git checkout jverhuls/Ardupilot-ibus-telemetry-feature-branch
M	modules/ChibiOS
Note: checking out 'jverhuls/Ardupilot-ibus-telemetry-feature-branch'.

You are in 'detached HEAD' state. You can look around, make experimental
changes and commit them, and you can discard any commits you make in this
state without impacting any branches by performing another checkout.

If you want to create a new branch to retain commits you create, you may
do so (now or later) by using -b with the checkout command again. Example:

  git checkout -b <new-branch-name>

HEAD is now at 8affe3512b Merge branch 'Ardupilot-ibus-telemetry-feature-branch' of https://github.com/jverhuls/ardupilot into Ardupilot-ibus-telemetry-feature-branch
pbarker@bluebottle:~/rc/ardupilot((HEAD detached at jverhuls/Ardupilot-ibus-telemetry-feature-branch))$ git merge master
Already up to date.
pbarker@bluebottle:~/rc/ardupilot((HEAD detached at jverhuls/Ardupilot-ibus-telemetry-feature-branch))$ git diff remotes/origin/master >/tmp/some.patch
pbarker@bluebottle:~/rc/ardupilot((HEAD detached at jverhuls/Ardupilot-ibus-telemetry-feature-branch))$ git checkout -b tmp remotes/origin/master
Previous HEAD position was 8affe3512b Merge branch 'Ardupilot-ibus-telemetry-feature-branch' of https://github.com/jverhuls/ardupilot into Ardupilot-ibus-telemetry-feature-branch
Branch 'tmp' set up to track remote branch 'master' from 'origin'.
Switched to a new branch 'tmp'
pbarker@bluebottle:~/rc/ardupilot(tmp)$ patch -p1 </tmp/some.patch
patching file Tools/ardupilotwaf/ardupilotwaf.py
patching file libraries/AP_IBus_Telem/AP_IBus_Telem.cpp
patching file libraries/AP_IBus_Telem/AP_IBus_Telem.h
patching file libraries/AP_SerialManager/AP_SerialManager.h
patching file libraries/AP_Vehicle/AP_Vehicle.cpp
patching file libraries/AP_Vehicle/AP_Vehicle.h
pbarker@bluebottle:~/rc/ardupilot(tmp)$ git add libraries/AP_IBus_Telem/AP_IBus_Telem.cpp libraries/AP_IBus_Telem/AP_IBus_Telem.h
pbarker@bluebottle:~/rc/ardupilot(tmp)$ git commit -C 23cd811bf40fdeeddb156997b09f64d4630e0b68 -a
[tmp e242e7ffeb] AP_IBus_Telem: optimize code - adding feature flag - styling
 Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
 Date: Sat Feb 13 08:04:01 2021 +0100
 6 files changed, 342 insertions(+)
 create mode 100644 libraries/AP_IBus_Telem/AP_IBus_Telem.cpp
 create mode 100644 libraries/AP_IBus_Telem/AP_IBus_Telem.h
pbarker@bluebottle:~/rc/ardupilot(tmp)$ git commit --amend -m "add support for IBus protocol"
[tmp fec51a0ead] add support for IBus protocol
 Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
 Date: Sat Feb 13 08:04:01 2021 +0100
 6 files changed, 342 insertions(+)
 create mode 100644 libraries/AP_IBus_Telem/AP_IBus_Telem.cpp
 create mode 100644 libraries/AP_IBus_Telem/AP_IBus_Telem.h
pbarker@bluebottle:~/rc/ardupilot(tmp)$ git show
commit fec51a0ead20ff4f33961d890eb75b27584e008b (HEAD -> tmp)
Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
Date:   Sat Feb 13 08:04:01 2021 +0100

    add support for IBus protocol

diff --git a/Tools/ardupilotwaf/ardupilotwaf.py b/Tools/ardupilotwaf/ardupilotwaf.py
index fb2d2be0be..f83eaa244f 100644
--- a/Tools/ardupilotwaf/ardupilotwaf.py
+++ b/Tools/ardupilotwaf/ardupilotwaf.py
@@ -64,6 +64,7 @@ COMMON_VEHICLE_DEPENDENT_LIBRARIES = [
     'AP_Button',
     'AP_ICEngine',
     'AP_Frsky_Telem',
+    'AP_IBus_Telem',
     'AP_FlashStorage',
     'AP_Relay',
     'AP_ServoRelayEvents',
diff --git a/libraries/AP_IBus_Telem/AP_IBus_Telem.cpp b/libraries/AP_IBus_Telem/AP_IBus_Telem.cpp
new file mode 100644
index 0000000000..e635f34383
pbarker@bluebottle:~/rc/ardupilot(tmp)$ ./Tools/gittools/git-subsystems-split 
rm: cannot remove '/tmp/tmp.7ZKk8CmyeI': No such file or directory
Reseting changes in order to add files separately...
Adding and committing subsystems...
[tmp 1aca27019e] AP_IBus_Telem: add support for IBus protocol
 Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
 2 files changed, 326 insertions(+)
 create mode 100644 libraries/AP_IBus_Telem/AP_IBus_Telem.cpp
 create mode 100644 libraries/AP_IBus_Telem/AP_IBus_Telem.h

[tmp a1f6a43e7e] AP_SerialManager: add support for IBus protocol
 Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
 1 file changed, 1 insertion(+)

[tmp c68b693d2b] AP_Vehicle: add support for IBus protocol
 Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
 2 files changed, 14 insertions(+)

[tmp 198d28d8b1] Tools: add support for IBus protocol
 Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
 1 file changed, 1 insertion(+)

pbarker@bluebottle:~/rc/ardupilot(tmp)$ git log
commit 198d28d8b1faa37adb912d720b56d38929707bc9 (HEAD -> tmp)
Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
Date:   Sun Feb 14 21:52:41 2021 +1100

    Tools: add support for IBus protocol

commit c68b693d2babf977200f4fb096cc023d32cf7336
Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
Date:   Sun Feb 14 21:52:40 2021 +1100

    AP_Vehicle: add support for IBus protocol

commit a1f6a43e7e9ceb88083b62376fed5de9b3996490
Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
Date:   Sun Feb 14 21:52:40 2021 +1100

    AP_SerialManager: add support for IBus protocol

commit 1aca27019ef7c7f49626ce1dc7ba61c8a7226de0
Author: Jan Verhulst <jan.verhulst@external.atlascopco.com>
Date:   Sun Feb 14 21:52:40 2021 +1100

    AP_IBus_Telem: add support for IBus protocol

commit 1b74058842954e7257c6475f21042be2102fe768 (origin/master, origin/HEAD, pr/ibus, master)
Author: Peter Barker <pbarker@barker.dropbear.id.au>
Date:   Wed Feb 10 12:16:43 2021 +1100
pbarker@bluebottle:~/rc/ardupilot(tmp)$ git push github tmp
Counting objects: 22, done.
Delta compression using up to 4 threads.
Compressing objects: 100% (18/18), done.
Writing objects: 100% (22/22), 6.65 KiB | 6.65 MiB/s, done.
Total 22 (delta 14), reused 6 (delta 4)
remote: Resolving deltas: 100% (14/14), completed with 9 local objects.
remote: 
remote: Create a pull request for 'tmp' on GitHub by visiting:
remote:      https://github.com/peterbarker/ardupilot/pull/new/tmp
remote: 
To ssh://github.com/peterbarker/ardupilot.git
 * [new branch]            tmp -> tmp
pbarker@bluebottle:~/rc/ardupilot(tmp)$ 
```

If you're happy with that branch (https://github.com/ArduPilot/ardupilot/compare/master...peterbarker:tmp?expand=1) then I'll push it over your github branch.